create_flashcards/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cached-property==1.5.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
chevron==0.14.0
distro==1.9.0
exceptiongroup==1.2.2
frozendict==2.4.4
genanki==0.13.1
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
jiter==0.5.0
oauthlib==3.2.2
openai==1.41.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
